Fixed to cleanup logic in XObjectsTests library.

+++ XObjectsTests/SetupAndCleanup.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
+using Microsoft.Schemas.Developer.MsBuild;
 using NUnit.Framework;
 
 namespace Xml.Schema.Linq.Tests
@@ -21,6 +22,9 @@ namespace Xml.Schema.Linq.Tests
             Console.SetOut(RedirectedOutput);
         }
 
+        /// <summary>
+        /// Cleans up any schema folders from previous runs that may not have been cleaned up due to errors.
+        /// </summary>
         [OneTimeTearDown]
         public static void Cleanup()
         {
@@ -30,13 +34,21 @@ namespace Xml.Schema.Linq.Tests
                     dir.Delete(true);
                 }
             }
-
+            
             const string guidRegex = @"(?im)[{(]?[0-9A-F]{8}[-]?(?:[0-9A-F]{4}[-]?){3}[0-9A-F]{12}[)}]?";
             var currentDir = new DirectoryInfo(@".").EnumerateDirectories("*", SearchOption.TopDirectoryOnly);
             var dirStrings = currentDir
                 .Where(d => d.Name.StartsWith("schemas_") && Regex.Match(d.Name, guidRegex).Success);
             foreach (var dir in dirStrings) {
-                dir.Delete(true);
+                dir.Refresh();
+                if (dir.Exists) {
+                    dir.Delete(true);
+                }
+            }
+
+            var mergedConfig = @".\merged.config";
+            if (File.Exists(mergedConfig)) {
+                File.Delete(mergedConfig);
             }
         }
     }
